+++ packages/runtime/container-runtime/src/gc/gcTelemetry.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ITelemetryLogger } from "@fluidframework/common-definitions";
+import { ITelemetryGenericEvent, ITelemetryLogger } from "@fluidframework/common-definitions";
 import { IGarbageCollectionData } from "@fluidframework/runtime-definitions";
 import { packagePathToTelemetryProperty } from "@fluidframework/runtime-utils";
 import {
@@ -17,8 +17,11 @@ import {
 	GCNodeType,
 	UnreferencedState,
 	IGarbageCollectorConfigs,
+	disableTombstoneKey,
+	throwOnTombstoneUsageKey,
+	throwOnTombstoneLoadKey,
+	runSweepKey,
 } from "./gcDefinitions";
-import { sendGCUnexpectedUsageEvent } from "./gcHelpers";
 import { UnreferencedStateTracker } from "./gcUnreferencedStateTracker";
 
 type NodeUsageType = "Changed" | "Loaded" | "Revived";
@@ -94,6 +97,7 @@ export class GCTelemetryTracker {
 		nodeType: GCNodeType,
 		usageType: NodeUsageType,
 		nodeStateTracker: UnreferencedStateTracker,
+		uniqueEventId: string,
 	) {
 		if (nodeStateTracker.state === UnreferencedState.Active) {
 			return false;
@@ -108,11 +112,9 @@ export class GCTelemetryTracker {
 			return false;
 		}
 
-		const uniqueEventId = `${nodeStateTracker.state}-${nodeId}-${usageType}`;
 		if (this.loggedUnreferencedEvents.has(uniqueEventId)) {
 			return false;
 		}
-		this.loggedUnreferencedEvents.add(uniqueEventId);
 		return true;
 	}
 
@@ -128,6 +130,9 @@ export class GCTelemetryTracker {
 			return;
 		}
 
+		// We log these events once per event per node. A unique id is generated by joining node state (inactive / sweep ready),
+		// node's id and usage (loaded / changed / revived).
+		const uniqueEventId = `${nodeStateTracker.state}-${nodeUsageProps.nodeId}-${nodeUsageProps.usageType}`;
 		const nodeType = this.getNodeType(nodeUsageProps.nodeId);
 		if (
 			!this.shouldLogNonActiveEvent(
@@ -135,10 +140,14 @@ export class GCTelemetryTracker {
 				nodeType,
 				nodeUsageProps.usageType,
 				nodeStateTracker,
+				uniqueEventId,
 			)
 		) {
 			return;
 		}
+
+		// Add the unique event id so that we don't generate a log for this event again in this session..
+		this.loggedUnreferencedEvents.add(uniqueEventId);
 
 		const state = nodeStateTracker.state;
 		const { usageType, currentReferenceTimestampMs, packagePath, ...propsToLog } =
@@ -351,4 +360,30 @@ export class GCTelemetryTracker {
 			});
 		}
 	}
+}
+
+/**
+ * Consolidates info / logic for logging when we encounter unexpected usage of GC'd objects. For example, when a
+ * tombstoned or deleted object is loaded.
+ */
+export function sendGCUnexpectedUsageEvent(
+	mc: MonitoringContext,
+	event: ITelemetryGenericEvent & {
+		category: "error" | "generic";
+		gcTombstoneEnforcementAllowed: boolean | undefined;
+	},
+	packagePath: readonly string[] | undefined,
+	error?: unknown,
+) {
+	event.pkg = packagePathToTelemetryProperty(packagePath);
+	event.tombstoneFlags = JSON.stringify({
+		DisableTombstone: mc.config.getBoolean(disableTombstoneKey),
+		ThrowOnTombstoneUsage: mc.config.getBoolean(throwOnTombstoneUsageKey),
+		ThrowOnTombstoneLoad: mc.config.getBoolean(throwOnTombstoneLoadKey),
+	});
+	event.sweepFlags = JSON.stringify({
+		EnableSweepFlag: mc.config.getBoolean(runSweepKey),
+	});
+
+	mc.logger.sendTelemetryEvent(event, error);
 }
